Refactor ProfileDropdown
