+++ static/js/components/ProfileDropdown.jsx
@@ -1,4 +1,5 @@
-import React, { Component } from 'react';
+import React, { useRef } from 'react';
+import { useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 import Dropdown, { DropdownTrigger, DropdownContent } from 'react-simple-dropdown';
@@ -7,80 +8,72 @@ import styles from "./ProfileDropdown.css";
 import Responsive from "./Responsive";
 
 
-class ProfileDropdown extends Component {
-  constructor(props) {
-    super(props);
-    this._collapse = this._collapse.bind(this);
-  }
+const ProfileDropdown = (props) => {
+  const profile = useSelector(state => state.profile);
+  const dropdown = useRef(null);
 
-  _collapse() {
-    this.dropdown.hide();
-  }
+  const collapseDropdown = () => {
+    dropdown.current.hide();
+  };
 
-  render() {
-    return (
-      <Responsive
-        desktopStyle={styles.profileDesktop}
-        mobileStyle={styles.profileMobile}
-      >
+  return (
+    <Responsive
+      desktopStyle={styles.profileDesktop}
+      mobileStyle={styles.profileMobile}
+    >
 
-        <Dropdown ref={(el) => { this.dropdown = el; }}>
+      <Dropdown ref={dropdown}>
 
-          <DropdownTrigger>
-            { this.props.profile.username }
-            {' '}
-            &nbsp;▾
-          </DropdownTrigger>
+        <DropdownTrigger>
+          { profile.username }
+          {' '}
+          &nbsp;▾
+        </DropdownTrigger>
 
-          <DropdownContent>
+        <DropdownContent>
 
-            <Link to="/profile">
-              <div className={styles.entry} onClick={this._collapse}>
-                Profile
-              </div>
-            </Link>
+          <Link to="/profile">
+            <div className={styles.entry} onClick={collapseDropdown}>
+              Profile
+            </div>
+          </Link>
 
-            <div className={styles.rule} />
+          <div className={styles.rule} />
 
-            <Link to="/groups">
-              <div className={styles.entry} onClick={this._collapse}>
-                Groups
-              </div>
-            </Link>
+          <Link to="/groups">
+            <div className={styles.entry} onClick={collapseDropdown}>
+              Groups
+            </div>
+          </Link>
 
-            <div className={styles.rule} />
+          <div className={styles.rule} />
 
-            <a href="https://github.com/skyportal/skyportal/issues/new">
-              <div className={styles.entry} onClick={this._collapse}>
-                File an issue
-              </div>
-            </a>
+          <a href="https://github.com/skyportal/skyportal/issues/new">
+            <div className={styles.entry} onClick={collapseDropdown}>
+              File an issue
+            </div>
+          </a>
 
-            <a href="https://github.com/skyportal/skyportal">
-              <div className={styles.entry} onClick={this._collapse}>
-                Help
-              </div>
-            </a>
+          <a href="https://github.com/skyportal/skyportal">
+            <div className={styles.entry} onClick={collapseDropdown}>
+              Help
+            </div>
+          </a>
 
-            <div className={styles.rule} />
+          <div className={styles.rule} />
 
-            <a href="/logout">
-              <div className={styles.entry} onClick={this._collapse}>
-                Sign out
-              </div>
-            </a>
+          <a href="/logout">
+            <div className={styles.entry} onClick={collapseDropdown}>
+              Sign out
+            </div>
+          </a>
 
-          </DropdownContent>
+        </DropdownContent>
 
-        </Dropdown>
+      </Dropdown>
 
-      </Responsive>
-    );
-  }
-}
-
-ProfileDropdown.propTypes = {
-  profile: PropTypes.object.isRequired
+    </Responsive>
+  );
 };
 
 export default ProfileDropdown;
